+++ BackEnd/version2/services/basket_services/read.go
@@ -10,9 +10,6 @@ import (
 	"github.com/walle692/D0018E/BackEnd/version2/utils/basket"
 )
 
-// GetBasket returns the contents of the user's basket
-// gets the basket id, then looks at all basketitems connected to that basket id
-// and returns the quantity and price + product name for each item.
 func GetBasket(c *gin.Context) {
 	session := sessions.Default(c)
 	userIDStr := session.Get(global.UserID)
@@ -20,7 +17,7 @@ func GetBasket(c *gin.Context) {
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 	}
-	basketItems, err := basket.GetUserBasket(userID)
+	basketItems, err := basket.GetByUserID(userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get users basket"})
 	}
