+++ BackEnd/version2/utils/tests/basketfunctions_test.go
@@ -28,7 +28,7 @@ func TestBasketToOrder(t *testing.T) {
 		qty := p1qty + 1
 		basketItemID := test_setup.SeedBasketItem(t, ctx, pool, basketID, p1, qty)
 
-		orderID, err := basket.ConvertBasketToOrder(basketID)
+		orderID, err := basket.Checkout(buyerID)
 		require.Error(t, err)
 		require.Zero(t, orderID)
 		//Check that order item wasn't deleted
@@ -49,7 +49,7 @@ func TestBasketToOrder(t *testing.T) {
 		price, qty := 100.0, 10
 		basketItemID := test_setup.SeedBasketItem(t, ctx, pool, basketID, p1, qty)
 
-		orderID, err := basket.ConvertBasketToOrder(basketID)
+		orderID, err := basket.Checkout(buyerID)
 		require.NoError(t, err)
 		require.NotZero(t, orderID)
 		//Test that basket item is removed from table
@@ -89,7 +89,7 @@ func TestBasketToOrder(t *testing.T) {
 		price_2, qty_2 := 200.0, 20
 		basketItemID_2 := test_setup.SeedBasketItem(t, ctx, pool, basketID, p2, qty_2)
 
-		orderID, err := basket.ConvertBasketToOrder(basketID)
+		orderID, err := basket.Checkout(buyerID)
 		require.NoError(t, err)
 		require.NotZero(t, orderID)
 		//Test that basket item is removed from table
@@ -129,7 +129,7 @@ func TestBasketToOrder(t *testing.T) {
 
 		pool.Exec(ctx, `UPDATE myschema.products SET active = FALSE WHERE product_id=$1`, p1)
 
-		orderID, err := basket.ConvertBasketToOrder(basketID)
+		orderID, err := basket.Checkout(buyerID)
 		require.Error(t, err)
 		t.Cleanup(func() { deleteNewOrder(t, ctx, pool, orderID) })
